+++ runloop.c
@@ -4143,16 +4143,6 @@ static bool event_init_content(
       RARCH_LOG("[SRAM]: %s\n",
             msg_hash_to_str(MSG_SKIPPING_SRAM_LOAD));
 
-#ifdef HAVE_BSV_MOVIE
-   bsv_movie_deinit(input_st);
-   if (bsv_movie_init(input_st))
-   {
-      /* Set granularity upon success */
-      configuration_set_uint(settings,
-            settings->uints.rewind_granularity, 1);
-   }
-#endif
-
 /*
    Since the operations are asynchronous we can't
    guarantee users will not use auto_load_state to cheat on
@@ -4163,15 +4153,42 @@ static bool event_init_content(
 #ifdef HAVE_CHEEVOS
    if (!cheevos_enable || !cheevos_hardcore_mode_enable)
 #endif
-#ifdef HAVE_BSV_MOVIE
-     if (!input_st->bsv_movie_state_handle)
-#endif
    {
-      if (runloop_st->entry_state_slot && !command_event_load_entry_state(settings))
-         runloop_st->entry_state_slot = 0;
-      if (!runloop_st->entry_state_slot && settings->bools.savestate_auto_load)
-         command_event_load_auto_state();
+#ifdef HAVE_BSV_MOVIE
+     // ignore entry state if we're doing bsv playback (we do want it
+     // for bsv recording though)
+     if (!(input_st->bsv_movie_state.flags & BSV_FLAG_MOVIE_START_PLAYBACK))
+#endif
+      {
+         if (runloop_st->entry_state_slot && !command_event_load_entry_state(settings))
+         {
+            // loading the state failed, reset entry slot
+            runloop_st->entry_state_slot = 0;
+         }
+      }
+#ifdef HAVE_BSV_MOVIE
+     // ignore autoload state if we're doing bsv playback or recording
+     if (!(input_st->bsv_movie_state.flags & (BSV_FLAG_MOVIE_START_RECORDING | BSV_FLAG_MOVIE_START_PLAYBACK)))
+#endif
+      {
+        if (!runloop_st->entry_state_slot && settings->bools.savestate_auto_load)
+          command_event_load_auto_state();
+      }
    }
+
+#ifdef HAVE_BSV_MOVIE
+   movie_stop(input_st);
+   if(input_st->bsv_movie_state.flags & BSV_FLAG_MOVIE_START_RECORDING)
+   {
+     configuration_set_uint(settings, settings->uints.rewind_granularity, 1);
+     movie_start_record(input_st, input_st->bsv_movie_state.movie_start_path);
+   }
+   else if(input_st->bsv_movie_state.flags & BSV_FLAG_MOVIE_START_PLAYBACK)
+   {
+     configuration_set_uint(settings, settings->uints.rewind_granularity, 1);
+     movie_start_playback(input_st, input_st->bsv_movie_state.movie_start_path);
+   }
+#endif
 
    command_event(CMD_EVENT_NETPLAY_INIT, NULL);
